src/server/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pyyaml>=6.0
aiofiles>=23.0.0
httpx>=0.27.0
